Replace endlessly recursive `from_u8` implementation
The `from_u8` macro implements `From<u8>`, but it uses `.into()` internally, which is automatically derived from `From<u8>` !
Could use mem::transmute for most of them, but there are a couple of edge cases like Ucssel, Ucastp. I think a manual implementation is a bit cleaner/safer.

Not sure why this worked in the first place, perhaps those registers are never read in the examples.