Fix clock selection bug
The inline assembly in `fll_on` and `fll_off` is subtly wrong - instead of masking the 6th bit of the status register (i.e masking SR with a value of 64), as written it instead grabs the contents of the address 64 bytes after the current instruction and uses *that* value as the SR mask. Very occasionally this actually works which makes the problem appear inconsistent, as the effects vary based on the compiler version, the memory layout, optimisation settings, the phase of the moon, etc. This is why the nop's had an effect - they could affect which value was used to set/clear the SR bits. With this fixed the nop's can be removed.

Also, in the `SmclkDisabled` version of .freeze() the wait states are set after configuring the clocks instead of before (like the SmclkDefined version does).