Make I2C transaction, write_read, compliant with ehal-1.0
The previous implementation of write_read was not compliant with the contract (had a stop pattern in the middle of it), and the new 1.0 `transaction` didn't match the contract either. Both failed around sending start and stops between operations.

Make the write, read functions take bools to determine if they produce start and stop patterns during transmission so they can be used as part of `write_read` and `transaction`

Remove unimplemented ArbitrationLost I2C error