Implement LPM
When writing an incorrect password to PMMCTL0 to lock it again, writing to the whole PMMCTL0 register (as opposed to just the PMMCTL0 high byte) causes a system reset.

The RTC only needs to be set once (survives LPM3.5 poweroff), so provide an unsafe method to enter LPM3.5 that doesn't require the RTC to be provided in a configured state.