Fix GPIO output pin state being modified during initialisation
The state of a GPIO pin is conserved during a system reset (watchdog timeout, etc).
But the `PxOUT` register was always being written to during initialisation, so if the pin was an output the state of the pin was changed during initialisation.
Instead of using `pxout_wr()`, use a call to `pxout_set()` and `pxout_clr()` to give us the option to leave the value unchanged.

Previous behaviour:
If `Input<Pullup>` set `PxOUT`
If `Input<Pulldown>` clear `PxOUT`
If `Input<Floating>` don't care, clear `PxOUT`
If `Output` clear `PxOUT` (bug!)

New behaviour:
If `Input<Pullup>` set `PxOUT`
If `Input<Pulldown>` clear `PxOUT`
If `Input<Floating>` don't care, don't change
If `Output` don't change